Update HamburgerMenu docs
If you don't explicitly set a different foreground color for the children of HamburgerMenu, they'll inherit from their parent, which in this example is "white." If this can be confusing in basic cases (like a page that displays a text box) because if the page background is *also* white (which by default, it is) it'll seem like the children aren't rendering. 
Setting "foreground" fixes this.