Update Directory.Build.targets to include Windows.winmd
When a target framework of `net47` is detected, the Windows.winmd is added as a reference based on target platform version/min version. Reference is added as series of hint paths so the most recent SDK available for that version is selected. If no TPV/TPMV is specified, the defaults are used from Directory.Build.props.