Provide reason for IWebView to be public
In scenarios where the toolbox or designer is not used, an instance of WebView can be created from an instance of WebViewControlProcess. We provide a wrapper around the WinRT type and a series of extension methods for WinForms and WPF that create controls with WinForms/WPF types (e.g. System.Drawing.Rectangle for WinForms, System.Windows.Rect for WPF). The unit test host has been updated to use the public factory method instead of internals (like a customer).