Support navigation methods after InitializeComponent
When a named WebView is placed into a WPF application, it is the expectation of the developer that Navigate(Uri), Navigate(String), and NavigateToString(String) may be called and navigations occur.

Problem: The control has yet to finish initialization, specific the method BuildCoreWindow has not yet been called and we do not have handles from HwndHost. Creation of the underlying WWAHost process is asynchronous to allow the remainder of the WPF UI to render. Navigation is now executed on a Dispatcher thread asynchronously on Background priority, waiting for the initialization to be completed. Navigate(String) and Navigate(Uri) use the Source dependency property, which is read after initialization and set on the underlying control already.

All other methods/properties invoked after InitializeComponent should not return any values.