Replace packages.config for modules with directive
To use long path module in Cake, a packages.config file was checked in. Instead of having two separate files with various elements for build, the `#module` directive in Cake is used instead. The `build.ps1` is altered to bootstrap Cake prior to execution via `.\cake.exe build.cake --bootstrap`.