Fix paths for batch file build shortcuts
The batch files required the prompt to be in the directory, using relative file path `build.ps1`. This has been changed to use the COMSPEC macro for current execution location, enabling the file to be executed from any location and be able to find the correct location. The path has also been quoted so any clones to paths with spaces are executed as expected.