Update WPF NavigateToString
In some situations a race condition froze the dispatcher queue and prevented any messages to be processed, causing the WPF UI to be unresponsive until the window was moved or resized (forcing new messages onto the queue). The cause for this was in the implementation to asynchronously push an action to the dispatcher on the Background priority but on the UI thread--the background threads were waiting on the UI and UI waiting on the background threads, causing the deadlock.

The implementation has been updated to make use of the ManualResetEvent for signaling initialization completed; however, it has been updated to:
 - yield thread time back if the object is not signaled in 200ms
 - asynchronously push a no-op frame at ContextIdle priority to force queue to flush
 - variable containing web view instance marked as volatile to ensure most recent version is read

This has reliability proven to not have deadlocks in various debugging and runtime scenarios, and improves performance while using the NavigateToString method.