Update Layout to send Client RECT instead of Bounds
When the layout is updated the HWND used the raw Bounds and did not account for the offsets required within the window (e.g. title bar, additional controls, etc.). In a typical control, the `DisplayRectangle` is the interior canvas of the control, and in a scrolling control the `DisplayRectangle` would be larger than the `ClientRectangle`. However, the difference in the two RECTs is abstracted from us in WebView, so we only need to synchronize the `ClientRectangle` and WebView will handle scrolling based on the new view port.