Add try/catch to UnsubscribeProcessExited
Under normal tear down, we want to unsubscribe from events to permit GC; however, when the underlying process hosting web view (Win32WebViewHost.exe) is terminated unexpectedly, the Win32 API returns a HRESULT 0x8007042B (The process terminated unexpectedly). Since the unsubscription is best effort, we catch all exceptions and attempt to continue. In the event of a catastrophic failure during the unsubscription process, we will attempt to continue execution with raising the event handler.

Resolves #2145