Add support for Partition property
The property was introduced in a later version of the SDK. The value is gated on a check on `ApiInformation.IsPropertyPresent` to verify the property is available on the type. If the property is not present, the wrapper type holds on to the value and it is simply dropped when creating the WinRT type with the process options.