Add additional mitigation for path traversal attack
During the URI conversion, perform an additional check for "dot-dot-slash" path traversal attack by throwing an exception when any invalid char is detected in a path. The implementation of `BuildStream` should already mitigate this, but adding a check anyway.

Code has been updated so that the test stream resolver uses the production version. A suppression for SCS0018 is also added warning of the path traversal vulnerability since it is mitigated by both checking for invalid characters in the path and further upstream in the `BuildStream` method.