Add script for installing Windows SDK from ISO
Script will download the ISO image from a URI, write out the cache information and use HTTP features for cache. If the file is out of date, a new file is downloaded in its place. If the file does not exist, is out of date, or it cannot be determined from the HTTP header, a new file is downloaded.  After the file is downloaded the ISO image is mounted, and the installation proceeds. Once finished the ISO is dismounted.