Remove `global::` prefix for WinRT types
The ability to access a member in the global namespace is useful when a member might be hidden by another entity of the same name.

Example
```
class Foo
{
  public class System { }

  const int Console = 0;

  static void Main()
  {
    Console.WriteLine("Bar");
  }
}
```
The `Console.WriteLine` throws an error because it accesses Foo.Console, which is a constant. Using `System.Console` does no better, since it accesses `Foo.System` class, which does not have a static `WriteLine` method. The only way to get it to compile is to use `global::System.Console.WriteLine("Bar")`; however, it is not required in these instances.