Add OS build check when initializing the control
According to report in #2411, there is an issue when initializing the control in some circumstances that cause thread to not complete initializing the underlying control. This is reported to be fixed by @MichaelVach by replacing the initialization on L97 with a simpler await.

Updated to fetch the OS build and check for RS4 range. If not RS4, don't use RS4 work around.

NOTE: the issue does not reproduce for me, and the code works in both code paths on OS builds 17134, 17744, and 18227.