Prevent _xamlSource from leaking
This field is only needed to surface the content through the `WindowsXamlHost.Child` property and does not otherwise need to be exposed to inherited classes when the property will suffice. To maintain the same API surface as the WPF control, a new API `SetContent` was added to take the `FrameworkElement`. The implementation of `SetContent(FrameworkElement)` is identical to `SetContent(UIElement)` other than the parameter type.

CC @ryalanms