Add entry for WebViewCompatible in manifest
When referencing the WebView NuGet, only WebView is shown in the toolbox despite also having another UI control WebViewCompatible also available. XAML entry could be made by hand, but not by dragging off the toolbox.

Update adds an entry for WebViewCompatible which permits it to show in the toolbox. Resolves #2482