Add `AddPreLoadedScript(string)` to WebView
- Need documentation for how to use AddPreLoadedScript.
- API appears to be non-functional at this point in time.
Update Win32 WebView API due to change in 17723

New to 17723 is a change of name for the AddPreLoadedScript. It is now named AddInitializeScript. The existing method has been marked obsolete with directions to use the new method name. The old method will be removed in a future release of the toolkit.

Cherry picked from 08b9b1453511350852f08e445a9aae20455ccb17