Revert "Added [StackTraceHidden] attribute"
This reverts commit f70ae0cdb29055d0cf1637f9c43163af9900524d.
Unfortunately the attribute is actually matched by type and not just by name like the others. We should reintroduce this once either https://github.com/dotnet/runtime/issues/29681 is done, or when the runtime switches to a name matching for the internal attribute, so we can just use a local copy here.