Fixed bug in TileBuilder small tile content
The enum flags were incorrectly specified such that when adding content to Medium/Wide/Large tiles, it also added the content to small tiles, which was resulting in the small weather tile not appearing correctly.

Note that the small weather tile preview is still missing the temperature (must be bug in NotificationsVisualizerLibrary), but when pinned to Start it looks correct!