Optimize UITests Shared project's item includes
Remove redundant 'Exclude' as bin/obj doesn't exist in the Shared project.

We use None to include items to be visible under the Shared Project
in VS IDE. But having both None and Compile/Page includes for the
same files will impact Build and IDE perf.

So, Only include None when Compile/Page is already active.