Fix issue with ColorPicker not working with SwitchPresenter changes
Since optimizing SwitchPresenter, the delayed loading being in another control's template was causing an issue where the cases weren't available to evaluate when SwitchPresenter was ready. Since we no longer listened to all downstream changes in SwitchPresenter this meant we never evaluated once the parent control was loaded. Added a couple of cases to ensure that once we're fully loaded we can evaluate again to ensure we pick a case. This also covers a case where if somehow the entire collection gets swapped out later.

Added some extra checks ahead of the evaluate cases to break out early in case we don't need to update the Visual Tree.