Fix ILT0010 warnings
Code ILT0010: Could not find an assembly referenced by other assembly

The following warnings surfaced due to the missing assembly.
Since, the assembly is not found, any referenced code got hit with these warnings

Code ILT0005: The type from an assembly was not included in compilation, but was referenced in a method.
Code ILT0003: The method will always throw an exception due to the missing method in an assembly.