Merge '~ProgressBarValue' classes
Meld 'BindableProgressBarValue' class into 'AdaptiveProgressBarValue' class
The difference between them is subtle and is platform specific. So, melding
the changes from one class into the other shouldn't be an issue.

Although this is a Breaking Change, it doesn't change any functional behavior.
It's refactored to be a drop-in replacement. So, just a global text replace is enough.