Wrong formatting applied by .NET Tool
Both VS IDE Code Clean-up and dotnet-format applies this formatting.
Only Code labels and Code within pre-processor blocks are affected.

The reason may be that the indentation wasn't read/written properly
by the formatter within or around those code blocks.

NOTE: We could ignore this commit but we keep it as a reference point
      for the VS C# Editor and .NET format team to diagnose this issue.