Remove invalid binding - add failing tests to show problems for #4749
Think we maybe used to bind to the text directly, but there's no Text property directly on the TokenizingTextBoxItem, so this binding is meaningless. It doesn't effect the behavior of the textbox in practical usage, but somehow was doing something which was masking the problem for us to be able to detect within a test case.
Removing it to reduce confusion.

The text sync between the parent TokenizingTextBox collection of items (and the current edit) and the item is maintained through code-behind with text changing events. Though work is being done to resolve issues in this sync process. See issue https://github.com/CommunityToolkit/WindowsCommunityToolkit/issues/4749